Feature/employee document (#37)
* --Added employee documents fileld---

* --Add emploee documents field  --

* --Added employee document browse file--9

* --Added constant string--

+++ KHRMS/Controllers/EmployeeAttendanceController.cs
@@ -1,4 +1,5 @@
-﻿using KHRMS.Core;
+﻿using Azure;
+using KHRMS.Core;
 using KHRMS.Infrastructure;
 using KHRMS.Services;
 using Microsoft.AspNetCore.Mvc;
@@ -14,7 +15,7 @@ namespace KHRMS
     {
         private readonly IEmployeeAttendanceService _attendanceService = employeeAttendanceService;
 
-    
+
 
         [HttpGet("GetAll")]
         public async Task<IActionResult> GetAll()
@@ -23,11 +24,11 @@ namespace KHRMS
             return Ok(new ApiResponse<IEnumerable<EmployeeAttendance>>
             {
                 StatusCode = (int)HttpStatusCode.OK,
-                Message = "Employee attendance records retrieved successfully.",
+                Message = ApiMessageConstant.AllAttendanceRequestsFound,
                 Data = attendances
             });
         }
-       
+
 
         [HttpGet]
         [Route("GetByEmployeeId/{employeeId}")]
@@ -52,7 +53,7 @@ namespace KHRMS
 
             return Ok(response);
         }
-      
+
 
         [HttpPost]
         [Route("AddEmployeeAttendanceRequest")]
@@ -76,7 +77,7 @@ namespace KHRMS
             };
             return Ok(response);
         }
- 
+
         [HttpPut]
         [Route("UpdateEmployeeAttendanceRequest")]
         public async Task<IActionResult> UpdateEmployeeAttendanceRequest(EmployeeAttendance attendance)
@@ -100,24 +101,26 @@ namespace KHRMS
             return Ok(response);
 
         }
-       
-        //extra
+
         [HttpDelete]
         [Route("DeleteEmployeeAttendanceRequest/{id}")]
         public async Task<IActionResult> DeleteEmployeeAttendanceRequest(long id)
         {
+         
+            var document = await _attendanceService.GetByIdAsync(id);
+            if (document == null)
+                return NotFound();
+
             await _attendanceService.DeleteAsync(id);
-            var response = new ApiResponse<bool>
+            return Ok(new ApiResponse<bool>
             {
                 StatusCode = (int)HttpStatusCode.OK,
                 Message = ApiMessageConstant.AttendanceRequestDeleted,
                 Data = true
-            };
-            return Ok(response);
+            });
+
         }
     }
-
-
 
 }
